test: avoid core imports during acceptance generation
