+++ packages/extension/src/extension/config/excludePatterns.ts
@@ -1,0 +1,17 @@
+/**
+ * Default exclude patterns for file discovery.
+ * These patterns avoid analyzing build artifacts, dependencies, and non-source files.
+ */
+export const DEFAULT_EXCLUDE_PATTERNS: readonly string[] = [
+  '**/node_modules/**',
+  '**/dist/**',
+  '**/build/**',
+  '**/out/**',
+  '**/.git/**',
+  '**/.codegraphy/**',
+  '**/.turbo/**',
+  '**/coverage/**',
+  '**/*.min.js',
+  '**/*.bundle.js',
+  '**/*.map',
+];
